Added updating to registration by national coach.

+++ src/main/java/com/sumo/server/Database/RegistrationAndWeightData/CompetitorRegistrationByNationalTeamCoach/CompetitorRegistrationByNationalTeamCoachRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CompetitorRegistrationByNationalTeamCoachRepository extends JpaRepository<CompetitorRegistrationByNationalTeamCoach, Integer> {
-
+public interface CompetitorRegistrationByNationalTeamCoachRepository extends JpaRepository<CompetitorRegistrationByNationalTeamCoach, Long> {
 }
